mon/src/main/java/org/bordylek/mon/model/Chart.java
package org.bordylek.mon.model;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

@Document(collection = "monitor")
public class Chart {

    @Id
    private String id;

    private String name;
    private String serie;
    private String size;

    public Chart() {
    }

    public Chart(String name, String serie, String size) {
        this.name = name;
        this.serie = serie;
        this.size = size;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getSerie() {
        return serie;
    }

    public void setSerie(String serie) {
        this.serie = serie;
    }

    public String getSize() {
        return size;
    }

    public void setSize(String size) {
        this.size = size;
    }
}
